fixed summary function

+++ backend/routes/products.js
@@ -34,12 +34,54 @@ router.post("/", isAdmin, async (req, res) => {
   }
 });
 
+//GET PRODUCT
+
+router.get("/find/:id", async (req, res) => {
+  try {
+    const product = await Product.findById(req.params.id);
+    res.status(200).send(product);
+  } catch (error) {
+    res.status(500).send(error);
+  }
+});
+
 //DELETE
 
 router.delete("/:id", isAdmin, async (req, res) => {
   try {
-    await Product.findByIdAndDelete(req.params.id);
-    res.status(200).send("Product has been deleted...");
+    const product = await Product.findById(req.params.id);
+
+    if (!product) return res.status(404).send("product not found ...");
+    
+    if (product.image.public_id) {
+      const destroyResponse = await cloudinary.uploader.destroy(
+        product.image.public_id
+        );
+
+      if (destroyResponse) {
+        const deletedProduct = await Product.findByIdAndDelete(req.params.id);
+        res.status(200).send(deletedProduct);
+      }
+    } else {
+      console.log("Action terminated. Failed to deleted product image ...");
+    }
+  } catch (error) {
+    res.status(500).send(error);
+  }
+});
+
+//UPDATE
+
+router.put("/:id", isAdmin, async (req, res) => {
+  try {
+    const updatedProduct = await Product.findByIdAndUpdate(
+      req.params.id,
+      {
+        $set: req.body,
+      },
+      { new: true }
+    );
+    res.status(200).send(updatedProduct);
   } catch (error) {
     res.status(500).send(error);
   }
@@ -66,32 +108,5 @@ router.get("/", async (req, res) => {
   }
 });
 
-//GET PRODUCT
-
-router.get("/find/:id", async (req, res) => {
-  try {
-    const product = await Product.findById(req.params.id);
-    res.status(200).send(product);
-  } catch (error) {
-    res.status(500).send(error);
-  }
-});
-
-//UPDATE
-
-router.put("/:id", isAdmin, async (req, res) => {
-  try {
-    const updatedProduct = await Product.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: req.body,
-      },
-      { new: true }
-    );
-    res.status(200).send(updatedProduct);
-  } catch (error) {
-    res.status(500).send(error);
-  }
-});
 
 module.exports = router;
